+++ src/components/Header/Header.jsx
@@ -35,6 +35,7 @@ export default function Header({ onClick }) {
                 <li><Link to="/goalslist">{isEn ? "Goals" : "Цілі"}</Link></li>
                 <li><Link to="/calendar">{isEn ? "Calendar" : "Календар"}</Link></li>
                 <li><Link to="/help">{isEn ? "Instructions" : "Інструкція"}</Link></li>
+                <li><Link to="/notebook">{isEn ? "Notebook" : "Нотатки"}</Link></li>
                 <li><Link to="/settings">{isEn ? "Settings" : "Налаштування"}</Link></li>
             </ul>
             <div className='burger-menu' onClick={onClick}>