+++ src/components/Menu/Menu.jsx
@@ -17,6 +17,7 @@ export default function Menu({ isMenu, onClick }) {
                 <li><Link to="/projects">{isEn ? "Projects" : "Проєкти"}</Link></li>
                 <li><Link to="/goalslist">{isEn ? "Goals" : "Цілі"}</Link></li>
                 <li><Link to="/calendar">{isEn ? "Calendar" : "Календар"}</Link></li>
+                <li><Link to="/notebook">{isEn ? "Notebook" : "Нотатки"}</Link></li>
                 <li><Link to="/settings">{isEn ? "Settings" : "Налаштування"}</Link></li>
             </ul>
         </div>